server/requirements.txt
alembic==1.15.1
aniso8601==10.0.0
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
cloudinary==1.42.2
colorama==0.4.6
Flask==3.1.0
Flask-Admin==1.6.1
flask-babel==4.0.0
flask-cors==5.0.1
Flask-DebugToolbar==0.16.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.9
MarkupSafe==3.0.2
pillow==11.1.0
Pygments==2.19.1
PyJWT==2.9.0
PyMySQL==1.1.1
pyotp==2.9.0
python-dotenv==1.0.1
pytz==2025.1
qrcode==8.0
referencing==0.36.2
rpds-py==0.23.1
six==1.17.0
SQLAlchemy==2.0.39
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==2.3.0
